% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_doc.R
\docType{data}
\name{acetileno}
\alias{acetileno}
\title{Geração de energia}
\format{
Data frame com 16 linhas e 4 variáveis:
\itemize{
\item conversao: conversão de n-heptano para acetileno (\%)
\item temperatura: temperatura de reação (em °C)
\item razao: razão de H2 para n-heptano (razão molar)
\item tempo: tempo de contato (em segundos)
}
}
\description{
Exemplo: Dados referentes à porcentagem de conversão de n-heptano para acetileno e três variáveis explicativas. Estes são dados típicos de processos químicos para os quais uma superfície de resposta quadrática completa envolvendo todas as três variáveis regressoras é frequentemente considerada como um modelo inicial de investigação.
}
\references{
Montgomery, D. C., Peck, E. A., and Vining, G. G. (2012) Introduction to Linear Regression Analysis, 5th Edition, Boca Raton: Wiley.
}
\author{
Fábio N. Demarqui \email{fndemarqui@est.ufmg.br}
}
\keyword{datasets}
