% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{dados_simulados}
\alias{dados_simulados}
\title{Dados Simulados}
\format{
Objeto \code{data.frame} com 6 colunas:

\describe{
\item{y}{Variável resposta gerada a partir da partir da expressão \code{y = 3x1 + x2^2 - 80x3 + rnorm(1000, mean = 0, sd = 1)}}
\item{x1}{Variável explicativa 1 gerada através de \code{rnorm(1000, mean = 0, sd = 1)}}
\item{x2}{Variável explicativa 2 gerada através de \code{rnorm(1000, mean = 0, sd = 3)}}
\item{x3}{Variável explicativa 3 gerada através de \code{rnorm(1000, mean = 4, sd = 1)}}
\item{x4}{Variável 4 gerada através de \code{rnorm(1000, mean = 5, sd = 2)}}
\item{x5}{Variável 5 gerada através de \code{rnorm(1000, mean = 1, sd = 1)}}
}
}
\source{
Os dados foram gerados a partir de variáveis aleatórias normalmente distribuídas.
As variáveis explicativas x1, x2, x3, x4 e x5 foram geradas utilizando a função
\code{rnorm} com diferentes médias e desvios padrão. A variável resposta y
foi calculada como uma combinação linear das variáveis x1, x2 e x3, adicionando
um erro aleatório normalmente distribuído. A semente utilizada para gerar as
simulações foi \code{set.seed(123456)}.
}
\usage{
dados_simulados
}
\description{
Dados simulados para testar as funções do pacote.
}
\keyword{datasets}
