% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_doc.R
\docType{data}
\name{entregas}
\alias{entregas}
\title{Tempo de Entrega de Refrigerantes}
\format{
Data frame com 25 linhas e 3 variáveis:
\itemize{
\item tempo: tempo de entrega (em minutos)
\item caixas: número de caixas a serem entregues
\item distancia: distância percorrida (em pés)
}
}
\description{
Exemplo: Uma engarrafadora de refrigerantes está analisando as rotas de serviço das máquinas de venda automática em seu sistema de distribuição. A empresa está interessada em prever a quantidade de tempo necessária para a rota percorrida pelo motorista para atender as máquinas de venda automática em um ponto de venda. Esta atividade de serviço inclui estocar a máquina com produtos de bebidas e a realização de pequenas manutenções. O engenheiro industrial responsável pelo estudo sugeriu que duas variáveis mais importantes que afetam o tempo de entrega são o número de caixas de produto estocado e a distância percorrida pelo roteirista . Foram coletadas 25 observações sobre o tempo de entrega.
}
\references{
Montgomery, D. C., Peck, E. A., and Vining, G. G. (2012) Introduction to Linear Regression Analysis, 5th Edition, Boca Raton: Wiley.
}
\author{
Fábio N. Demarqui \email{fndemarqui@est.ufmg.br}
}
\keyword{datasets}
