% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_doc.R
\docType{data}
\name{estocagem}
\alias{estocagem}
\title{Tempo de estocagem de refrigerantes em prateleiras}
\format{
Data frame com 20 linhas e duas variáveis:
\itemize{
\item tempo: tempo gasto para o estoque da mercadoria em uma prateleira.
\item caixas: número de caixas do produto estocado (covariável).
}
}
\description{
Exemplo 2.28: Banco de dados com o tempo necessário para um comerciante estocar uma prateleira de supermercado com um refrigerante bem como o número de caixas do produto estocado.
}
\references{
Montgomery, D. C., Peck, E. A., and Vining, G. G. (2012) Introduction to Linear Regression Analysis, 5th Edition, Boca Raton: Wiley.
}
\author{
Fábio N. Demarqui \email{fndemarqui@est.ufmg.br}
}
\keyword{datasets}
