% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_doc.R
\docType{data}
\name{gas}
\alias{gas}
\title{Isolamento térmico de casa e consumo de gás}
\format{
Data frame com 56 linhas e 3 variáveis:
\itemize{
\item isolamento: variável indicadora (antes/depois) da instalação do isolamento  (covariável)
\item temperatura:  temperatura externa medida em graus Celsius (covariável)
\item consumo: consumo semanal de gás em metros cúbicos  (variável resposta)
}
}
\description{
Derek Whiteside, da Estação de Pesquisa de Edifícios do Reino Unido, registrou o consumo semanal de gás e a temperatura externa média em sua própria casa no sudeste da Inglaterra para duas estações de aquecimento, uma de 26 semanas antes e uma de 30 semanas após a instalação do isolamento térmico nas paredes da sua residência. O objetivo do exercício foi avaliar o efeito do isolamento térmico no consumo de gás.
}
\references{
\insertRef{MASS}{reglin}
}
\author{
Fábio N. Demarqui \email{fndemarqui@est.ufmg.br}
}
\keyword{datasets}
