% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gginfluence.R
\name{gginfluence}
\alias{gginfluence}
\title{ggplot version of influence diagnostic plots for an lm object}
\usage{
gginfluence(
  object,
  measure = c("leverage", "dfbetas", "cooksd", "dffits", "covratio"),
  ...
)
}
\arguments{
\item{object}{an object of class lm with the fitted linear model.}

\item{measure}{the desired influence diagnostic measure; available options are:
\itemize{
\item leverage
\item dfbetas
\item cooksd
\item dffits
\item covratio
}}

\item{...}{further arguments passed to other methods.}
}
\description{
ggplot version of influence diagnostic plots for an lm object
}
\author{
Fábio N. Demarqui
}
