% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graficos.R
\name{graficos_diagnostico}
\alias{graficos_diagnostico}
\title{Gráficos de Diagnóstico}
\usage{
graficos_diagnostico(preditos, residuos, df)
}
\arguments{
\item{preditos}{Valores preditos retornados pela função \code{rl}.}

\item{residuos}{Resíduos retornados pela função \code{rl}.}

\item{df}{Banco de dados no formato \code{data.frame} quw contém a variável resposta.}
}
\value{
Uma figura com 6 gráficos: Valores predidos x Valores observados; QQplot dos resíduos; Valores preditos x Resíduos; Histograma dos resíduos; Número da observação x Resíduos; e Boxplot dos resíduos.
}
\description{
Fornece gráficos para o diagnóstico de um modelo de regressão linear.
}
\examples{
pred = rl("y", c("x1", "x2", "x3"), dados_simulados)$ValoresPreditos
resid = rl("y", c("x1", "x2", "x3"), dados_simulados)$Residuos
graficos_diagnostico(pred, resid, dados_simulados)

dados_simulados$x2_quadrado = (dados_simulados$x2)^2
pred2 = rl("y", c("x1", "x2_quadrado", "x3"), dados_simulados)$ValoresPreditos
resid2 = rl("y", c("x1", "x2_quadrado", "x3"), dados_simulados)$Residuos
graficos_diagnostico(pred2, resid2, dados_simulados)

}
