% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_doc.R
\docType{data}
\name{moinho}
\alias{moinho}
\title{Geração de energia}
\format{
Data frame com 25 linhas e duas variáveis:
\itemize{
\item velocidade: velocidade do vento (covariável)
\item tensao: tensão de saída (variável resposta)
}
}
\description{
Exemplo: Um engenheiro de pesquisa está investigando o uso de um moinho de vento para gerar eletricidade. Ele coletou dados sobre a tensão de saída de seu moinho de vento e a velocidade do vento correspondente.
}
\references{
Montgomery, D. C., Peck, E. A., and Vining, G. G. (2012) Introduction to Linear Regression Analysis, 5th Edition, Boca Raton: Wiley.
}
\author{
Fábio N. Demarqui \email{fndemarqui@est.ufmg.br}
}
\keyword{datasets}
