% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_doc.R
\docType{data}
\name{moscas}
\alias{moscas}
\title{Longevidade de moscas-das-frutas}
\format{
Data frame com 124 linhas e 3 variáveis:
\itemize{
\item torax: comprimento do tórax de cada macho (covariável)
\item longevidade:  longevidade da mosca-da-fruta em dias (variável resposta)
\item grupo:  grupos de interesse para comparação (covariável)
}
}
\description{
Em um experimento 125 moscas-das-frutas que foram divididas aleatoriamente em 5 grupos de 25 cada. A resposta foi a longevidade da mosca-da-fruta em dias. Um grupo foi mantido solitário (A), enquanto outro foi mantido individualmente com uma fêmea virgem a cada dia (B). Outro grupo recebeu 8 fêmeas virgens por dia (C). Como controle adicional, o quarto (D) e quinto (E) grupos foram mantidos com uma ou oito fêmeas grávidas por dia. As moscas-das-frutas grávidas não acasalam. O comprimento do tórax de cada macho foi medido, pois isso era conhecido por afetar a longevidade. Uma observação no grupo D foi perdida.
}
\references{
\insertRef{faraway}{reglin}
}
\author{
Fábio N. Demarqui \email{fndemarqui@est.ufmg.br}
}
\keyword{datasets}
