% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_doc.R
\docType{data}
\name{ouricos}
\alias{ouricos}
\title{Tamanho de ouriços}
\format{
Data frame com 72 linhas e 3 variáveis:
\itemize{
\item regime: regime de alimentação: A, B ou C (covariável)
\item tamanho_inicial: tamanho do ouriço antes do início da dieta (covariável)
\item tamanho_final: tamanho do ouriço no final do experimento (variável resposta)
}
}
\source{
\href{https://www.tidymodels.org/start/models/}{tidymodels}
}
\description{
Um experimento foi realizado com o objetivo de determinar como três regimes de alimentação diferentes (A, B e C) afetam o tamanho de ouriços-do-mar ao longo do tempo. O tamanho inicial dos ouriços-do-mar no início do experimento provavelmente afeta o tamanho que eles crescem à medida que são alimentados, e portanto foi considerado uma variável de perturbação.
}
\author{
Fábio N. Demarqui \email{fndemarqui@est.ufmg.br}
}
\keyword{datasets}
