% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_doc.R
\docType{data}
\name{oxigenio}
\alias{oxigenio}
\title{Nível de colesterol no sangue}
\format{
Data frame com 20 linhas e duas variáveis:
\itemize{
\item percentual: percentual de hidrocarbonetos presentes no condensador principal (covariável)
\item pureza: purezao do oxigênio (variável resposta)
}
}
\description{
Exemplo: Um engenheiro está interessado em estudar a relação entre a pureza do oxigênio (y) produzido em um processo químico de destilação e o percentagem de hidrocarbonetos (x) presente no condensador principal de uma unidade de destilação.
}
\references{
Montgomery, D. C., Runger, G. C. (2015) Estatística Aplicada e Probabilidade para Engenheiros, 5ª. Edição, Rio de Janeiro: LTC.
}
\author{
Fábio N. Demarqui \email{fndemarqui@est.ufmg.br}
}
\keyword{datasets}
