% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{pad}
\alias{pad}
\title{Scaling of variables}
\usage{
pad(x, type = c("normal", "unit"))
}
\arguments{
\item{x}{a vector containing the values to be scaled (normal and unity scales are available)}

\item{type}{type of scaling to be applied.}
}
\value{
a vector with the scaled variable
}
\description{
Scaling of variables
}
