% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pred.R
\name{predicao}
\alias{predicao}
\title{Predições para modelos de regressão linear}
\usage{
predicao(coeficientes_estimados, novos_dados)
}
\arguments{
\item{coeficientes_estimados}{Um vetor com os coeficientes estimados pela função \code{rl}.}

\item{novos_dados}{Banco de dados no formato \code{data.frame} com os novos valores para realizar as predições.}
}
\value{
Um \code{data.frame} com as predições para os novos valores.
}
\description{
Fornece predições para novos valores das variáveis explicativas usadas para ajustar um modelo de regressão linear.
}
\examples{
coef = rl("y", c("x1", "x2", "x3"), dados_simulados)$Coeficientes
df_novo = data.frame(x1 = c(1, 5), x2 = c(3, 4), x3 = c(2, 2))
predicao(coef, df_novo)

}
