% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/press.R
\name{press}
\alias{press}
\title{PRESS statistic}
\usage{
press(object, ...)
}
\arguments{
\item{object}{one or more objects of class lm.}

\item{...}{further arguments passed to or from other methods.}
}
\value{
the PRESS statistic when a single model is passed to the function; otherwise, a data.frame with the PRESS statistic values is returned.
}
\description{
Function to compute PRESS statistic for one or more models.
}
\examples{
\donttest{
library(reglin)
data(entregas)

fit0 <- lm(tempo ~ 1, data = entregas)
fit1 <- lm(tempo ~ caixas, data = entregas)
fit2 <- lm(tempo ~ distancia, data = entregas)
fit3 <- lm(tempo ~ caixas + distancia, data = entregas)
fit4 <- lm(tempo ~ caixas*distancia, data = entregas)

press(fit1)
press(fit1, fit2, fit3, fit4)
}
}
