% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R2.R
\name{R2}
\alias{R2}
\title{Coefficient of determination (R2)}
\usage{
R2(model)
}
\arguments{
\item{model}{an object of class lm or aov with the fitted linear model.}
}
\value{
the coefficient of determination associated with the fitted model.
}
\description{
Coefficient of determination (R2)
}
