% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R2.R
\name{R2adj}
\alias{R2adj}
\title{Adjusted Coefficient of determination (R2adj)}
\usage{
R2adj(model)
}
\arguments{
\item{model}{an object of class lm or aov with the fitted linear model.}
}
\value{
the coefficient of determination associated with the fitted model.
}
\description{
Adjusted Coefficient of determination (R2adj)
}
