% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reglin-package.R
\name{reglin-package}
\alias{reglin-package}
\alias{reglin}
\title{Pacote Reglin - Regressão Linear.}
\description{
Funções e bancos de dados para a disciplina de Análise de Regressão.
}
\details{
_PACKAGE
}
\references{
Montgomery, D. C., Peck, E. A, and Vining, G. G, (2016). Introduction to Linear Regression Analysis, 5ª. Edição, New York: Wiley.

Montgomery, D. C., Runger, G. C. (2015) Estatística Aplicada e Probabilidade para Engenheiros, 5ª. Edição, Rio de Janeiro: LTC.
}
