% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_doc.R
\docType{data}
\name{resistencia}
\alias{resistencia}
\title{Resistência de papel}
\format{
Data frame com 19 linhas e duas variáveis:
\itemize{
\item concentracao: percentual de madeira nobre no lote de celulose
\item resistencia: resistencia do papel
}
}
\description{
Exemplo: Dados relativos à resistência do papel kraft e ao percentual de madeira nobre no lote de celulose a partir do qual o papel foi produzido. O diagrama de dispersão desses dados, aliado ao conhecimento do processo de produção, sugere que um modelo quadrático pode descrever adequadamente a relação entre resistência à tração e concentração de madeira nobre.
}
\references{
Montgomery, D. C., Peck, E. A., and Vining, G. G. (2012) Introduction to Linear Regression Analysis, 5th Edition, Boca Raton: Wiley.
}
\author{
Fábio N. Demarqui \email{fndemarqui@est.ufmg.br}
}
\keyword{datasets}
