% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_doc.R
\docType{data}
\name{restaurantes}
\alias{restaurantes}
\title{Efeito de propaganda sobre as receitas de restaurantes}
\format{
Data frame com 3 linhas e 3 variáveis:
\itemize{
\item receita: receita média mensal do restaurante (variável resposta)
\item propaganda: gasto anual com propaganda (covariável)
\item id: indicadora de grupo de restaurantes
}
}
\description{
Exemplo: Um estudo foi realizado para avaliar o impacto do gasto anual com propagana sobre o rendimento médio mensal da venda de alimentos de 30 restaurantes.
}
\references{
Montgomery, D. C., Peck, E. A., and Vining, G. G. (2012) Introduction to Linear Regression Analysis, 5th Edition, Boca Raton: Wiley.
}
\author{
Fábio N. Demarqui \email{fndemarqui@est.ufmg.br}
}
\keyword{datasets}
