% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rl.R
\name{rl}
\alias{rl}
\title{Regressão Linear}
\usage{
rl(y, X, df)
}
\arguments{
\item{y}{Variável resposta da classe \code{character}}

\item{X}{Vetor de \code{character} com as variáveis explicativas}

\item{df}{Banco de dados no formato \code{data.frame}}
}
\value{
Uma lista contendo um vetor com os coeficientes, um vetor com os valores preditos e um vetor com os resíduos do modelo ajustado.
}
\description{
Fornece os coeficientes, os valores preditos e os resíduos de uma regressão linear.
}
\examples{
rl("y", c("x1", "x2", "x3"), dados_simulados)

dados_simulados$x2_quadrado = (dados_simulados$x2)^2
rl("y", c("x1", "x2_quadrado", "x3"), dados_simulados)

}
