% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{se.lm}
\alias{se.lm}
\title{Estimated standard errors for the regression coefficients}
\usage{
\method{se}{lm}(object, ...)
}
\arguments{
\item{object}{an object of the class lm}

\item{...}{further arguments passed to or from other methods}
}
\value{
the standard errors associated with the regression coefficients.
}
\description{
Estimated standard errors for the regression coefficients
}
