% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{standardization}
\alias{standardization}
\title{Variable standardization}
\usage{
standardization(x, type = c("normal", "length"))
}
\arguments{
\item{x}{a vector containing the values to be scaled (unit normal scaling and unit length scales)}

\item{type}{type of scaling to be applied.}
}
\value{
a vector with the standardized variable
}
\description{
Implementation of the unit normal and unit length scaling methods to produce standardized regression coefficients.
}
