% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tab_anova.R
\name{tab_anova}
\alias{tab_anova}
\title{ANOVA table for linear models}
\usage{
tab_anova(model)
}
\arguments{
\item{model}{an object of class lm or aov with the fitted linear model.}
}
\value{
the ANOVA table decomposed into model and error sources of variation.
}
\description{
ANOVA table for linear models
}
