% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_doc.R
\docType{data}
\name{tensao}
\alias{tensao}
\title{Queda de tensão da bateria em um motor de míssil guiado}
\format{
Data frame com 41 linhas e duas variáveis:
\itemize{
\item tempo: tempo após o lançamento (em segundos)
\item tensao: queda de tensão (Volt)
}
}
\description{
Exemplo: Dados referentes à queda de tensão da bateria em um motor de míssil guiado observada durante o voo do míssil. O diagrama de dispersão sugere que a queda de tensão se comporta de maneira diferente em diferentes segmentos de tempo e, portanto, modelaremos os dados com um spline cúbico usando dois nós em t1 = 6,5 e t2 = 13 segundos após o lançamento, respectivamente.
}
\references{
Montgomery, D. C., Peck, E. A., and Vining, G. G. (2012) Introduction to Linear Regression Analysis, 5th Edition, Boca Raton: Wiley.
}
\author{
Fábio N. Demarqui \email{fndemarqui@est.ufmg.br}
}
\keyword{datasets}
