% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testResiduals.R
\name{testResiduals}
\alias{testResiduals}
\title{Residuals Tests for Linear Regression Models}
\usage{
testResiduals(object, ...)
}
\arguments{
\item{object}{an object of class lm with the fitted linear model.}

\item{...}{further arguments passed to other methods.}
}
\description{
Residuals Tests for Linear Regression Models
}
\author{
Fábio N. Demarqui
}
