% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_doc.R
\docType{data}
\name{wdi}
\alias{wdi}
\title{PIB per capita e expectativa de vida de 211 países}
\format{
Data frame com 211 linhas e duas variáveis:
\itemize{
\item pib: PIB per capita, ou Produto Interno Bruto por habitante; o PIB per capita é um indicador econômico que reflete a riqueza média produzida por cada pessoa em um determinado local, como um país ou região, e é calculado dividindo o Produto Interno Bruto total pelo número de habitantes.
\item expectativa: expectativa de vida (em anos)
}
}
\description{
Conjunto de dados contendo o PIB per capita e a expectativa de vida de 211 países.
}
\references{
Izbicki, R. e Santos, T. M. dos. Aprendizado de máquina: uma abordagem estatística. 1ª edição. 2020. 272 páginas. ISBN: 978-65-00-02410-4.
}
\author{
Fábio N. Demarqui \email{fndemarqui@est.ufmg.br}
}
\keyword{datasets}
