% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R2.R
\name{R2adj}
\alias{R2adj}
\title{Adjusted Coefficient of determination (R2adj)}
\usage{
R2adj(object, ...)
}
\arguments{
\item{object}{object of class lm, usually, a result of a call to lm.}

\item{...}{other objects of class lm, usually, a result of a call to lm.}
}
\value{
the coefficient of determination associated with the fitted model.
}
\description{
Adjusted Coefficient of determination (R2adj)
}
